Add files via upload
Help file includes the grep command used. I will eventually put this into a bash script to enable the bash script to run the file and then run the grep command automatically. We can do the same thing with the python script to properly format the data tables we will need. 

The example input and output files are included in this. I will include more enrichments of UO2 and gad. I will also add more comments to the card provided.